PR review comment fixes
