+++ libs/services/designer-client-services/src/lib/base/index.ts
@@ -1,0 +1,2 @@
+export { BaseOperationManifestService, foreachOperationInfo } from './operationManifest';
+export type { BaseOperationManifestServiceOptions } from './operationManifest';
