+++ libs/services/designer-client-services/src/lib/base/manifests/foreach.ts
