+++ libs/services/designer-client-services/src/lib/base/manifests/query.ts
