+++ libs/services/designer-client-services/src/lib/consumption/index.ts
@@ -1,2 +1,2 @@
 export { ConsumptionSearchService } from './search';
-export { ConsumptionOperationManifestService } from './operationmanifest';
+export { ConsumptionOperationManifestService } from './operationManifest';
