+++ libs/services/designer-client-services/src/lib/consumption/manifests/liquid.ts
