+++ libs/services/designer-client-services/src/lib/standard/index.ts
@@ -1,8 +1,7 @@
 export { StandardConnectionService, IApiHubServiceDetails } from './connection';
 export { StandardConnectorService } from './connector';
 export { StandardGatewayService } from './gateway';
-export { StandardOperationManifestService, foreachOperationInfo } from './operationmanifest';
+export { StandardOperationManifestService } from './operationmanifest';
 export { StandardOAuthService } from './oAuth';
-export type { StandardOperationManifestServiceOptions } from './operationmanifest';
 export { StandardSearchService } from './search';
-export { frequencyValues } from './manifests/schedule';
+export { frequencyValues } from '../base/manifests/schedule';
